Add support for asymmetric loss probabilities
The default loss probability model (`prob`) assumes that probabilities
are the same for both link directions.  Add a new model
(`prob_asymmetric`) to support different probabilities depending on
traffic direction.  This is specified in the configuration file as:

```
(...)
model:
{
        type = "prob_asymmetric";

        default_prob = 1.0;
        links = (
                (0, 2, 0.800000),
                (2, 0, 0.000000)
                (0, 1, 0.500000)
        );
};
```

In the above example, traffic originated at node 0 destined for node 2 would be
applied a loss probality of 0.8.  Traffic in the opposite direction would have
a null loss probability.

If only one direction is listed in the `links` section, the default probability
`default_prob` will be applied to traffic in the opposite direction.
So, in the above example, traffic from 0 to 1 would have a 0.5 loss
probability and 1 to 0 would have 1.0.

Signed-off-by: Javier Cardona <jcardona@fb.com>