implement a packet filter [wip]
For some use cases we would like to drop certain frames based
on their contents rather than, or in addition to, packet error
rates.

This adds a prototype of that kind of functionality.  At present,
the filter language is largely unspecified but it is expected
that something like (e)BPF could be used for that.